leaks ok

+++ parsing/input_handler.c
@@ -6,7 +6,7 @@
 /*   By: pmeimoun <pmeimoun@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 16:51:02 by pmeimoun          #+#    #+#             */
-/*   Updated: 2025/10/18 17:43:21 by pmeimoun         ###   ########.fr       */
+/*   Updated: 2025/10/20 11:28:50 by pmeimoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,37 +25,50 @@ static int	handle_syntax_errors(char **split_input)
 	return (0);
 }
 
+static int	execute_pipeline(t_command *commands, t_export *export)
+{
+	t_pipex	*pipex;
+
+	pipex = malloc(sizeof(t_pipex));
+	if (!pipex || !init_pipex(pipex, commands))
+	{
+		free(pipex);
+		free_commands(commands);
+		return (0);
+	}
+	child_process(commands, pipex, export);
+	wait_child();
+	unlink("temp");
+	free(pipex);
+	free_commands(commands);
+	return (1);
+}
+
 static int	handle_tokens(char **split_input, t_export *export)
 {
 	t_token		*token_list;
 	t_command	*commands;
-	t_pipex		*pipex;
 
 	token_list = tokenizer(split_input, export->env);
-	free_split(split_input);
-	if (token_list)
+	if (!token_list)
+		return (0);
+	assign_filename_types(token_list);
+	if (!expand_tokens(token_list, export->env))
 	{
-		assign_filename_types(token_list);
-		if (!expand_tokens(token_list, export->env))
-			return(0);
-		commands = parser(token_list);
 		free_tokens(token_list);
-		printf("%d\n", commands->args[0][0]);
-		pipex = malloc(sizeof(t_pipex));
-		if (!init_pipex(pipex, commands))
-			return (0);
-		child_process(commands, pipex, export);
-		wait_child();
-		unlink("temp");
-		free(pipex);
-		free_commands(commands);
+		return (0);
 	}
-	return (1);
+	commands = parser(token_list);
+	free_tokens(token_list);
+	if (!commands)
+		return (0);
+	return (execute_pipeline(commands, export));
 }
 
 int	handle_input(char *input, t_export *export)
 {
 	char	**split_input;
+	int		ret;
 
 	if (ft_strlen(input) == 0)
 	{
@@ -74,5 +87,9 @@ int	handle_input(char *input, t_export *export)
 		return (1);
 	if (handle_syntax_errors(split_input))
 		return (1);
-	return (handle_tokens(split_input, export));
+	ret = handle_tokens(split_input, export);
+	free_split(split_input);
+	if (!ret)
+		return (1);
+	return (1);
 }
